requirements.txt
absl-py==1.3.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
Automat==20.2.0
autopep8==2.0.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
blis==0.7.7
cachetools==5.2.0
catalogue==2.0.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cloud-init==22.4.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cycler==0.11.0
cymem==2.0.6
dbus-python==1.2.18
debugpy==1.6.0
decorator==4.4.2
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
entrypoints==0.4
executing==0.8.3
fastai==2.7.10
fastbook==0.0.24
fastcore==1.5.27
fastdownload==0.0.5
fastjsonschema==2.15.3
fastprogress==1.0.2
flatbuffers==22.12.6
fonttools==4.33.3
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20
grpcio==1.51.1
h5py==3.7.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
imageio==2.23.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.6.4
incremental==21.3.0
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.11.0
keyring==23.5.0
kiwisolver==1.4.2
langcodes==3.3.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==8.10.0
moviepy==1.0.3
murmurhash==1.0.7
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.11.0
notebook==6.4.11
numpy==1.22.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.1
pdf2image==1.16.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
preshed==3.0.6
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.6
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.8.2
Pygments==2.12.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-magic==0.4.24
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
Send2Trash==1.8.0
sentencepiece==0.1.96
service-identity==18.1.0
six==1.16.0
smart-open==5.2.1
sos==4.4
soupsieve==2.3.2.post1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
ssh-import-id==5.11
stack-data==0.2.0
systemd-python==234
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
terminado==0.13.3
thinc==8.0.15
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
Twisted==22.1.0
typer==0.4.1
typing_extensions==4.2.0
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.9
wadllib==1.3.6
wasabi==0.9.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.6.0
wrapt==1.14.1
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
